src/BentoML/bentoml-requirements.txt
opencv-python==4.8.0.74
pandas==2.1.2
Pillow==10.0.1
torch
torchaudio
torchvision
ultralytics==8.0.207
bentoml[io-json]==1.1.10
matplotlib>=3.2.2
numpy>=1.18.5
protobuf<=3.20.1
requests>=2.23.0
